+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.41] - 2026-04-28
+
+### Docs
+- Update README.md
+- Update TODO.md
+
+### Test
+- Update tests/test_doctor_orchestrator.py
+
+### Other
+- Update regres/doctor_cli.py
+- Update regres/doctor_orchestrator.py
+
 ## [0.1.40] - 2026-04-28
 
 ### Docs